fix(core): solve gh-pages url build issue
Despite create-react-app's instructions to set the homepage to your
production url in package.json, I was still getting 404 after
deploying to github pages.

I tried setting it to "." for relative paths as mentioned here:
https://create-react-app.dev/docs/deployment/#serving-the-same-build-from-different-paths

This worked, and had the side benefit of fixing the same issue while
running the production build locally.